# agent-verse-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentVerse",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tenants/signup": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Signup",
        "description": "Create a new tenant account and return the initial API key.",
        "operationId": "signup_tenants_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tenants/me/keys": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List Keys",
        "description": "List all API keys for the current tenant. Raw keys are never returned here.",
        "operationId": "list_keys_tenants_me_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Create Key",
        "description": "Create a new API key. The raw key is returned ONLY in this response.",
        "operationId": "create_key_tenants_me_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/me/keys/{key_id}": {
      "delete": {
        "tags": [
          "tenants"
        ],
        "summary": "Revoke Key",
        "description": "Revoke (deactivate) an API key owned by the current tenant.",
        "operationId": "revoke_key_tenants_me_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/me/keys/{key_id}/rotate": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Rotate Key",
        "description": "Rotate an API key: create a replacement, optionally revoke the original.\n\nThe newly created key's raw secret is returned **once** in this response.",
        "operationId": "rotate_key_tenants_me_keys__key_id__rotate_post",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/me/llm": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get Llm Config",
        "description": "Return the current LLM provider config for this tenant (key never exposed).",
        "operationId": "get_llm_config_tenants_me_llm_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tenants"
        ],
        "summary": "Set Llm Config",
        "description": "Configure the LLM provider for this tenant. The API key is stored encrypted.",
        "operationId": "set_llm_config_tenants_me_llm_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMProviderConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateKeyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateKeyRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMProviderConfig": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "Provider name: anthropic | openai | gemini | groq | together | azure | ollama"
          },
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key",
            "description": "API key (stored encrypted in vault)"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Base URL override (for Ollama / Azure / vLLM)"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "description": "Default model slug",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "provider",
          "api_key"
        ],
        "title": "LLMProviderConfig",
        "description": "LLM provider configuration for a tenant."
      },
...
      "RotateKeyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "default": "Rotated Key"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "revoke_old": {
            "type": "boolean",
            "title": "Revoke Old",
            "default": true
          }
        },
        "type": "object",
        "title": "RotateKeyRequest",
        "description": "Request body for key rotation."
      },
...
      "SignupRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "title": "SignupRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}